Add initiallyopen attribute to the <popup> element
This add an attribute called initiallyopen to the <popup> element.
When this attribute is present on page load, the first such element
will be shown by default.

See [1] for the conversation and Open-UI resolution.

This CL also re-points the explainer links for all popup tests
to the new Open-UI repo.

[1] https://github.com/WICG/open-ui/issues/311#issuecomment-820655960

Bug: 1168738
Change-Id: Ib95d12234f359d503dc168e6799dd2a17b0dc18e