[GridNG] Compute the correct offset for out of flow items that start or end in the middle of a range
Previously, if an out of flow item happened to be in the middle of a
range, its offset and size were computed incorrectly. In this change,
helpers are added to compute the extra offset needed. Two web tests were
added to validate this behaviour.

Additionally, we now store range related data instead of the set indices
for out of flow items, as in this case we need the first and last
spanned range, as well as the offsets of the tracks. With this
information, we can just compute the set indices on demand.

Bug: 1045599
Change-Id: I26c8e1027d89c601de8e81d3f31850e061423844